Version Packages (#38)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## svelte-podcast@0.6.0

### Minor Changes

-   1a95ddd: Updates dependencies
-   9232f4d: Improves theming with css variables
    Improves developer logs
    Fixes various package export issues
    Adds player components:

    -   Player: `<PlayerStack />`
    -   Player: `<PlayerWidget />`
    -   Utility: `<HeadlessTimeline />`

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>