Version Packages (#42)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## svelte-podcast@0.7.0

### Minor Changes

-   16366ec: improvements to headless components
-   415d083: switch to jsdoc from ts

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>