Version Packages (#45)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## svelte-podcast@0.8.0

### Minor Changes

-   d18b0c7: refactor audio element logic
-   29ef657: switch to jsdoc from ts
-   d18b0c7: rewrite docs with code samples

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>