Version Packages (#49)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## svelte-podcast@0.9.0

### Minor Changes

-   528457f: Restructures package exports
-   528457f: - improved export structure
    -   simplify audio methods
    -   improve typescript types
-   bba2421: Improve example components
-   bba2421: improve progress component style overrides
-   f42a00f: removes autoplay as it doesn't behave as users would expect

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>